Separate AbstractDummyPlatformController as a base class
