+++ src/test/java/org/hobbit/core/components/BenchmarkControllerTest.java
@@ -17,19 +17,17 @@
 package org.hobbit.core.components;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Random;
-import java.util.concurrent.Semaphore;
 
-import org.apache.commons.io.Charsets;
 import org.apache.commons.io.IOUtils;
 import org.apache.jena.rdf.model.ModelFactory;
 import org.hobbit.core.Commands;
 import org.hobbit.core.Constants;
 import org.hobbit.core.TestConstants;
+import org.hobbit.core.components.dummy.AbstractDummyPlatformController;
 import org.hobbit.core.components.dummy.DummyComponentExecutor;
 import org.hobbit.core.rabbit.RabbitMQUtils;
 import org.hobbit.vocab.HobbitExperiments;
@@ -43,7 +41,6 @@ import org.junit.runners.Parameterized.Parameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.rabbitmq.client.AMQP.BasicProperties;
 import com.rabbitmq.client.MessageProperties;
 
 @RunWith(Parameterized.class)
@@ -198,38 +195,19 @@ public class BenchmarkControllerTest extends AbstractBenchmarkController {
         sendResultModel(ModelFactory.createDefaultModel());
     }
 
-    protected static class DummyPlatformController extends AbstractCommandReceivingComponent {
+    protected static class DummyPlatformController extends AbstractDummyPlatformController {
 
         public List<DummyComponentExecutor> dataGenExecutors = new ArrayList<DummyComponentExecutor>();
         public List<Thread> dataGenThreads = new ArrayList<>();
         public List<DummyComponentExecutor> taskGenExecutors = new ArrayList<DummyComponentExecutor>();
         public List<Thread> taskGenThreads = new ArrayList<>();
         public Random random = new Random();
-        private boolean readyFlag = false;
 
         private String sessionId;
-        private Semaphore terminationMutex = new Semaphore(0);
 
         public DummyPlatformController(String sessionId) {
             super();
             this.sessionId = sessionId;
-        }
-
-        protected void handleCmd(byte bytes[], String replyTo) {
-            ByteBuffer buffer = ByteBuffer.wrap(bytes);
-            int idLength = buffer.getInt();
-            byte sessionIdBytes[] = new byte[idLength];
-            buffer.get(sessionIdBytes);
-            String sessionId = new String(sessionIdBytes, Charsets.UTF_8);
-            byte command = buffer.get();
-            byte remainingData[];
-            if (buffer.remaining() > 0) {
-                remainingData = new byte[buffer.remaining()];
-                buffer.get(remainingData);
-            } else {
-                remainingData = new byte[0];
-            }
-            receiveCommand(command, remainingData, sessionId, replyTo);
         }
 
         public void receiveCommand(byte command, byte[] data, String sessionId, String replyTo) {
@@ -332,46 +310,5 @@ public class BenchmarkControllerTest extends AbstractBenchmarkController {
                 }
             }
         }
-
-        public void waitForControllerBeingReady() throws InterruptedException {
-            while (!readyFlag) {
-                Thread.sleep(500);
-            }
-        }
-
-        public void sendToCmdQueue(String address, byte command, byte data[], BasicProperties props)
-                throws IOException {
-            byte sessionIdBytes[] = RabbitMQUtils.writeString(address);
-            // + 5 because 4 bytes for the session ID length and 1 byte for the
-            // command
-            int dataLength = sessionIdBytes.length + 5;
-            boolean attachData = (data != null) && (data.length > 0);
-            if (attachData) {
-                dataLength += data.length;
-            }
-            ByteBuffer buffer = ByteBuffer.allocate(dataLength);
-            buffer.putInt(sessionIdBytes.length);
-            buffer.put(sessionIdBytes);
-            buffer.put(command);
-            if (attachData) {
-                buffer.put(data);
-            }
-            cmdChannel.basicPublish(Constants.HOBBIT_COMMAND_EXCHANGE_NAME, "", props, buffer.array());
-        }
-
-        @Override
-        public void receiveCommand(byte command, byte[] data) {
-        }
-
-        @Override
-        public void run() throws Exception {
-            readyFlag = true;
-            terminationMutex.acquire();
-        }
-
-        public void terminate() {
-            terminationMutex.release();
-        }
-
     }
 }
